+++ docs/topics/release-notes.md
@@ -73,6 +73,7 @@ You can determine your currently installed version using `pip freeze`:
 
     Note: `AutoSchema.__init__` now ensures `manual_fields` is a list.
     Previously may have been stored internally as `None`.
+
 * Remove ulrparse compatability shim; use six instead [#5579][gh5579]
 * Drop compat wrapper for `TimeDelta.total_seconds()` [#5577][gh5577]
 * Clean up all whitespace throughout project [#5578][gh5578]